Implement local-only mode for B2 uploads
For testing purposes, it's useful to have access to a "dummy" version of
the B2 library to enable uploading/downloading/deleting/etc files
without having to have a B2 account.

This works by passing a local path to AuthorizeDummyAccount (instead of
using AuthorizeAccount). This new method will create the directory if it
doesn't exist, and modify the Auth struct to contain fields that tell
the other B2 utility methods whether or not the operation should be done
locally on the machine instead of uploading to an actual bucket.