Populate bucket ID with file ID when local-only is enabled
Since the B2 API requires the bucket ID to be used for single-chunk
downloads and the file ID to be used for multi-chunk uploads, both
should be populated with the same value (the file ID/name) when
local-only mode is enabled. This way the library can be used the same
with or without access to an actual B2 bucket.