Use int64 for all file size ops, make logging optional
This change enforces int64 instead of int for anything related to file
sizes, since 32-bit machines could run into issues with larger files
when using this library.

This also makes logging optional, with a new SetLogging function for the
Service struct to enable/disable logging as needed.

The calls that authorize the user's B2 account and return a Service
struct now return a pointer to the struct in order to allow modification
of the logging state without having to replace the existing Service
struct.